Request timeouts, expiration and some refactoring
Refactored Request to have the publishing just like ResponseWriter,
also added implementation for io.Writer, and WriteHeader.

Timeout will now set the Expiration field for the amqp.Publishing so
that amqp will have the correct TTL on the message. It will not send the
message to any consumers if the TTL has passed.