Client: Implement NotifyPublish and NotifyReturn
Client will wait for confirmation from amqp-server before continuing.
And in the case of requests where no response is wanted, not return
until the request has been confirmed.

If a message can not be routed or (if immediate) is set not be routed
immediately, the client will return an error.

Returns are always handled (as long as the immediate or mandatory flags
are set).

Confirmations can be controlled by the Client#WithPublishSettings or
Client#WithConfirmMode.

This commit also improves some error logging to give some more
information about the request/response.