Ensure new context for each request
This ensures that the user has more control over how to gracefully shut down any
running handlers.

This is more modeled after `net/http.Server` where a new Context is created for
each request. This fixes an issue where the inherited context will cancel stuff
that you might want more control over.

A pointer to the `Server` is added to the context and the user can now ask the
server if it's is shutting down. This is also modeled after how the
`net/http.Server` adds a pointer to the `Server` on the request context.