Use amqp.DefaultDialer
This commit removes the old default dialer and uses the one exposed in
amqp which does exactly the same thing as the old internal
dialer. With that we can also remove the DialConf since the dialer in
amqp accepts a timeout duration as an argument setting the dial and
handshake timouts.