Bindings: Move queue declaration settings to the binding
This aligns more with how AMQP works and allows for more flexibility. It
is now possible to set durability, auto-delete, auto-ack etc on a
per-binding basis.

Connections are now named.

The PrefetchCount is moved to the binding. This allows for different
prefetch counts for different bindings.

Exclusive consumers are no longer supported for the server. This would
make the server fail when starting multiple instances. Exclusive
consumers are still used for client reply-to queues.

The FanoutBinding convenience function is removed. Fanout exchanges are
a bit weird and it is probably better to use a topic binding instead.

The Immediate flag was removed in RabbitMQ 3 and was never actually
supported by this library.