Bindings: Move queue declaration settings to the binding
Options for the queue and consumer configuration on the bindings are now
moved to the Binding config struct. This enable the user to have
different options for different handlers. This aligns more with how AMQP
generally. We also default to using qourum-queues when creating
bindings.

The PrefetchCount configuration is also moved to the binding. This
allows for different prefetch counts for different bindings.

Exchanges are now declared as a seperate step due to the fact that
RabbitMQ comes with all the default exchanges ready to be used by
default. Those exchanges have now gotten constants. Most often, users
will not need to declare any exchanges at all.

The Mandatory flag is now set on the request instead of on the client as
a whole.

The Immediate flag was removed in RabbitMQ 3 and was never really
supported by this library.

ResponseWriter no longer has the option to set the Mandatory flag.
Mandatory would close the server connection if the client is no longer
running and the reply-to queue removed.

Connections are now named. Allowing them to be identified in the
management ui.

The FanoutBinding convenience function is removed. Fanout exchanges are
a bit weird and it is probably better to use a topic binding instead.