Fix issue when channel/connection is closed before NotifyClose
This fixes an issue where the connection or channel would be closed
before NotifyClose is setup. That can cause the monitorAndWait to block
even when a connection or channel is actually closed. This happens
because the consumers or publishers can close the channel due to an
error inside their respective goroutines.

This fix also avoids starting and leaking a bunch of goroutines inside
monitorAndWait.