Use proper context deadline to cancel requests
This enables a request to be canceled when it's context is canceled.
Perhaps due to an external timeout.

This also removes the chanSendWaitTime in favor of a context cancel.

When a user sets a shorter deadline with a cause on the context, that
cause will be returned by Send when the method returns.