Refactor OnConnected
Rename the Server.OnStarted and Client.OnStarted to
Server/Client.OnConnected and have the callback run when actually
connected. This is more useful since this enables the user to know when
the bindings and consumers are actually created.

Rename Client.runForever to Client.Connect. This enables the user to
pre-start the client to avoid a higher latency during the first call to
Send().