Change logging level for unknown request confirmation
After the changes in v5.0.0 where the confirms and response handling
where merged into one, we started to see "got confirmation of unknown
request" errors. It turns out that this is completly expected since it
is not guaranteed that the confirmation will be received before the
response. In the previous implementation the request would always wait
for the confirmation.

We can change the requestsMap to keep the request until confirmation
arrives, even after the response has been received. But this is more
overhead and not needed. During the 7 years of using this library in
production, we have never seen an unknown confirmation.

See commit 1551ff6d35ff23d9a98102794c6d5ad79e0b3d39 for where we
simplified the code and merged the confirms and response handling.

See https://www.rabbitmq.com/docs/confirms#publisher-confirms-latency
for information about the confirmation latency.